rename deprecated -> _deprecated
