+++ sncosmo/_deprecated.py
